pyproject.toml
[project]
name = "repository-reader"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "langchain-google-genai (>=2.1.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
